avances en modelo de dominio, combate, fixed bug on roll.js

+++ TPC_Equipo-10/DomainModel/RolledAbility.cs
@@ -19,7 +19,7 @@ namespace DomainModel
 
         public int GetModifier()
         {
-            return (int)((rolledScore - 10) / 2);
+            return (rolledScore - 10) / 2;
         }
     }
 }
